Use bigint for order_number

+++ ddl-tpcds/bin_partitioned/web_returns.sql
@@ -17,7 +17,7 @@ create table web_returns
     wr_returning_addr_sk      int,
     wr_web_page_sk            int,
     wr_reason_sk              int,
-    wr_order_number           int,
+    wr_order_number           bigint,
     wr_return_quantity        int,
     wr_return_amt             float,
     wr_return_tax             float,
